+++ tests/unit/test_layer_factory.py
@@ -2,6 +2,7 @@ import numpy as np
 
 from uniplot.layer_factory import render_points
 from uniplot.options import Options, CharacterSet
+from uniplot.color import Color
 
 
 def test_ascii_characters_without_color():
@@ -14,7 +15,7 @@ def test_ascii_characters_without_color():
     assert "+" in matrix
 
 
-def test_ascii_characters_with_color():
+def test_ascii_characters_with_terminal_color():
     xs = [np.array([1, 2, 3])]
     ys = [np.array([2, 1, 4])]
     opts = Options(
@@ -23,10 +24,27 @@ def test_ascii_characters_with_color():
         x_max=5.0,
         y_min=0.0,
         y_max=5.0,
-        color=["red"],
+        color=[Color.from_terminal("red")],
     )
     matrix = render_points(xs, ys, opts)
     assert _wrap_in_red("+") in matrix
+
+
+def test_ascii_characters_with_rgb_color():
+    xs = [np.array([1, 2, 3])]
+    ys = [np.array([2, 1, 4])]
+    opts = Options(
+        character_set=CharacterSet.ASCII,
+        x_min=0.0,
+        x_max=5.0,
+        y_min=0.0,
+        y_max=5.0,
+        height=3,
+        width=3,
+        color=[Color.from_rgb(146, 255, 12)],
+    )
+    matrix = render_points(xs, ys, opts)
+    assert _wrap_in_lime_green("+") in matrix
 
 
 def test_block_characters_without_color():
@@ -40,7 +58,9 @@ def test_block_characters_without_color():
 def test_block_characters_with_color():
     xs = [np.array([0, 5])]
     ys = [np.array([0, 5])]
-    opts = Options(x_min=0.0, x_max=5.0, y_min=0.0, y_max=5.0, color=["red"])
+    opts = Options(
+        x_min=0.0, x_max=5.0, y_min=0.0, y_max=5.0, color=[Color.from_terminal("red")]
+    )
     matrix = render_points(xs, ys, opts)
     assert _wrap_in_red("▖") in matrix
 
@@ -93,7 +113,7 @@ def test_braille_characters_with_color():
         width=3,
         height=3,
         character_set=CharacterSet.BRAILLE,
-        color=["red"],
+        color=[Color.from_terminal("red")],
     )
     matrix = render_points(xs, ys, opts)
     print(matrix)
@@ -107,3 +127,7 @@ def test_braille_characters_with_color():
 
 def _wrap_in_red(char: str) -> str:
     return "\033[31m" + str(char) + "\033[0m"
+
+
+def _wrap_in_lime_green(char: str) -> str:
+    return "\033[38;2;146;255;12m" + str(char) + "\033[0m"